requirements.txt
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.6.1
matplotlib==3.10.1
seaborn==0.11.1
tensorflow==2.19.0
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
pytest==6.2.5
python-dotenv==0.19.0
PyYAML==6.0.2
joblib==1.4.2